requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
awsebcli==3.20.3
bcrypt==3.2.2
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
cryptography==37.0.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.17.0
fastapi==0.78.0
future==0.16.0
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==0.10.0
jsonschema==3.2.0
mongoengine==0.24.1
motor==3.0.0
odmantic==0.5.0
paramiko==2.11.0
passlib==1.7.4
pathspec==0.9.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
PyYAML==5.4.1
requests==2.26.0
rsa==4.8
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
starlette==0.19.1
starlette-context==0.3.4
termcolor==1.1.0
texttable==1.6.4
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.1.9
websocket-client==0.59.0
websockets==10.3
